Actualizar lnk_msp430f169.cmd, main.c y 48 más archivos...

+++ MSP430F169_DriverLib/DriverLib/DMA/Peripheral/RegisterDefines/Header/DMA_RegisterDefines_CH_CTL.h
@@ -31,11 +31,11 @@
  ************************************ 3 CH_CTL *********************************************
  ******************************************************************************************/
 /*-----------*/
-#define DMA_CH_CTL_R_REQ_BIT ((uint8_t) 0U)
+#define DMA_CH_CTL_R_REQ_BIT ((uint16_t) 0U)
 
-#define DMA_CH_CTL_REQ_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_REQ_NA ((uint8_t) 0x00U)
-#define DMA_CH_CTL_REQ_REQUEST ((uint8_t) 0x01U)
+#define DMA_CH_CTL_REQ_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_REQ_NA ((uint16_t) 0x00U)
+#define DMA_CH_CTL_REQ_REQUEST ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_REQ_MASK (DMA_CH_CTL_REQ_MASK << DMA_CH_CTL_R_REQ_BIT)
 #define DMA_CH_CTL_R_REQ_NA (DMA_CH_CTL_REQ_NA << DMA_CH_CTL_R_REQ_BIT)
@@ -43,11 +43,11 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_ABORT_BIT ((uint8_t) 1U)
+#define DMA_CH_CTL_R_ABORT_BIT ((uint16_t) 1U)
 
-#define DMA_CH_CTL_ABORT_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_ABORT_NA ((uint8_t) 0x00U)
-#define DMA_CH_CTL_ABORT_ABORTED ((uint8_t) 0x01U)
+#define DMA_CH_CTL_ABORT_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_ABORT_NA ((uint16_t) 0x00U)
+#define DMA_CH_CTL_ABORT_ABORTED ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_ABORT_MASK (DMA_CH_CTL_ABORT_MASK << DMA_CH_CTL_R_ABORT_BIT)
 #define DMA_CH_CTL_R_ABORT_NA (DMA_CH_CTL_ABORT_NA << DMA_CH_CTL_R_ABORT_BIT)
@@ -55,11 +55,11 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_IE_BIT ((uint8_t) 2U)
+#define DMA_CH_CTL_R_IE_BIT ((uint16_t) 2U)
 
-#define DMA_CH_CTL_IE_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_IE_DIS ((uint8_t) 0x00U)
-#define DMA_CH_CTL_IE_ENA ((uint8_t) 0x01U)
+#define DMA_CH_CTL_IE_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_IE_DIS ((uint16_t) 0x00U)
+#define DMA_CH_CTL_IE_ENA ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_IE_MASK (DMA_CH_CTL_IE_MASK << DMA_CH_CTL_R_IE_BIT)
 #define DMA_CH_CTL_R_IE_DIS (DMA_CH_CTL_IE_DIS << DMA_CH_CTL_R_IE_BIT)
@@ -67,11 +67,11 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_IFG_BIT ((uint8_t) 3U)
+#define DMA_CH_CTL_R_IFG_BIT ((uint16_t) 3U)
 
-#define DMA_CH_CTL_IFG_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_IFG_NOOCCUR ((uint8_t) 0x00U)
-#define DMA_CH_CTL_IFG_OCCUR ((uint8_t) 0x01U)
+#define DMA_CH_CTL_IFG_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_IFG_NOOCCUR ((uint16_t) 0x00U)
+#define DMA_CH_CTL_IFG_OCCUR ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_IFG_MASK (DMA_CH_CTL_IFG_MASK << DMA_CH_CTL_R_IFG_BIT)
 #define DMA_CH_CTL_R_IFG_NOOCCUR (DMA_CH_CTL_IFG_NOOCCUR << DMA_CH_CTL_R_IFG_BIT)
@@ -79,11 +79,11 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_EN_BIT ((uint8_t) 4U)
+#define DMA_CH_CTL_R_EN_BIT ((uint16_t) 4U)
 
-#define DMA_CH_CTL_EN_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_EN_DIS ((uint8_t) 0x00U)
-#define DMA_CH_CTL_EN_ENA ((uint8_t) 0x01U)
+#define DMA_CH_CTL_EN_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_EN_DIS ((uint16_t) 0x00U)
+#define DMA_CH_CTL_EN_ENA ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_EN_MASK (DMA_CH_CTL_EN_MASK << DMA_CH_CTL_R_EN_BIT)
 #define DMA_CH_CTL_R_EN_DIS (DMA_CH_CTL_EN_DIS << DMA_CH_CTL_R_EN_BIT)
@@ -91,11 +91,11 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_LEVEL_BIT ((uint8_t) 5U)
+#define DMA_CH_CTL_R_LEVEL_BIT ((uint16_t) 5U)
 
-#define DMA_CH_CTL_LEVEL_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_LEVEL_EDGE ((uint8_t) 0x00U)
-#define DMA_CH_CTL_LEVEL_LEVEL ((uint8_t) 0x01U)
+#define DMA_CH_CTL_LEVEL_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_LEVEL_EDGE ((uint16_t) 0x00U)
+#define DMA_CH_CTL_LEVEL_LEVEL ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_LEVEL_MASK (DMA_CH_CTL_LEVEL_MASK << DMA_CH_CTL_R_LEVEL_BIT)
 #define DMA_CH_CTL_R_LEVEL_EDGE (DMA_CH_CTL_LEVEL_EDGE << DMA_CH_CTL_R_LEVEL_BIT)
@@ -103,11 +103,11 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_SRCBYTE_BIT ((uint8_t) 6U)
+#define DMA_CH_CTL_R_SRCBYTE_BIT ((uint16_t) 6U)
 
-#define DMA_CH_CTL_SRCBYTE_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_SRCBYTE_WORD ((uint8_t) 0x00U)
-#define DMA_CH_CTL_SRCBYTE_BYTE ((uint8_t) 0x01U)
+#define DMA_CH_CTL_SRCBYTE_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_SRCBYTE_WORD ((uint16_t) 0x00U)
+#define DMA_CH_CTL_SRCBYTE_BYTE ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_SRCBYTE_MASK (DMA_CH_CTL_SRCBYTE_MASK << DMA_CH_CTL_R_SRCBYTE_BIT)
 #define DMA_CH_CTL_R_SRCBYTE_WORD (DMA_CH_CTL_SRCBYTE_WORD << DMA_CH_CTL_R_SRCBYTE_BIT)
@@ -115,11 +115,11 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_DSTBYTE_BIT ((uint8_t) 7U)
+#define DMA_CH_CTL_R_DSTBYTE_BIT ((uint16_t) 7U)
 
-#define DMA_CH_CTL_DSTBYTE_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_DSTBYTE_WORD ((uint8_t) 0x00U)
-#define DMA_CH_CTL_DSTBYTE_BYTE ((uint8_t) 0x01U)
+#define DMA_CH_CTL_DSTBYTE_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_DSTBYTE_WORD ((uint16_t) 0x00U)
+#define DMA_CH_CTL_DSTBYTE_BYTE ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_DSTBYTE_MASK (DMA_CH_CTL_DSTBYTE_MASK << DMA_CH_CTL_R_DSTBYTE_BIT)
 #define DMA_CH_CTL_R_DSTBYTE_WORD (DMA_CH_CTL_DSTBYTE_WORD << DMA_CH_CTL_R_DSTBYTE_BIT)
@@ -127,13 +127,13 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_SRCINCR_BIT ((uint8_t) 8U)
+#define DMA_CH_CTL_R_SRCINCR_BIT ((uint16_t) 8U)
 
-#define DMA_CH_CTL_SRCINCR_MASK ((uint8_t) 0x03U)
-#define DMA_CH_CTL_SRCINCR_UNCHANGED ((uint8_t) 0x00U)
-#define DMA_CH_CTL_SRCINCR_UNCHANGED1 ((uint8_t) 0x01U)
-#define DMA_CH_CTL_SRCINCR_DECREMENT ((uint8_t) 0x02U)
-#define DMA_CH_CTL_SRCINCR_INCREMENT ((uint8_t) 0x03U)
+#define DMA_CH_CTL_SRCINCR_MASK ((uint16_t) 0x03U)
+#define DMA_CH_CTL_SRCINCR_UNCHANGED ((uint16_t) 0x00U)
+#define DMA_CH_CTL_SRCINCR_UNCHANGED1 ((uint16_t) 0x01U)
+#define DMA_CH_CTL_SRCINCR_DECREMENT ((uint16_t) 0x02U)
+#define DMA_CH_CTL_SRCINCR_INCREMENT ((uint16_t) 0x03U)
 
 #define DMA_CH_CTL_R_SRCINCR_MASK (DMA_CH_CTL_SRCINCR_MASK << DMA_CH_CTL_R_SRCINCR_BIT)
 #define DMA_CH_CTL_R_SRCINCR_UNCHANGED (DMA_CH_CTL_SRCINCR_UNCHANGED << DMA_CH_CTL_R_SRCINCR_BIT)
@@ -143,13 +143,13 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_DSTINCR_BIT ((uint8_t) 10U)
+#define DMA_CH_CTL_R_DSTINCR_BIT ((uint16_t) 10U)
 
-#define DMA_CH_CTL_DSTINCR_MASK ((uint8_t) 0x03U)
-#define DMA_CH_CTL_DSTINCR_UNCHANGED ((uint8_t) 0x00U)
-#define DMA_CH_CTL_DSTINCR_UNCHANGED1 ((uint8_t) 0x01U)
-#define DMA_CH_CTL_DSTINCR_DECREMENT ((uint8_t) 0x02U)
-#define DMA_CH_CTL_DSTINCR_INCREMENT ((uint8_t) 0x03U)
+#define DMA_CH_CTL_DSTINCR_MASK ((uint16_t) 0x03U)
+#define DMA_CH_CTL_DSTINCR_UNCHANGED ((uint16_t) 0x00U)
+#define DMA_CH_CTL_DSTINCR_UNCHANGED1 ((uint16_t) 0x01U)
+#define DMA_CH_CTL_DSTINCR_DECREMENT ((uint16_t) 0x02U)
+#define DMA_CH_CTL_DSTINCR_INCREMENT ((uint16_t) 0x03U)
 
 #define DMA_CH_CTL_R_DSTINCR_MASK (DMA_CH_CTL_DSTINCR_MASK << DMA_CH_CTL_R_DSTINCR_BIT)
 #define DMA_CH_CTL_R_DSTINCR_UNCHANGED (DMA_CH_CTL_DSTINCR_UNCHANGED << DMA_CH_CTL_R_DSTINCR_BIT)
@@ -159,17 +159,17 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_DT_BIT ((uint8_t) 12U)
+#define DMA_CH_CTL_R_DT_BIT ((uint16_t) 12U)
 
-#define DMA_CH_CTL_DT_MASK ((uint8_t) 0x07U)
-#define DMA_CH_CTL_DT_SINGLE ((uint8_t) 0x00U)
-#define DMA_CH_CTL_DT_BLOCK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_DT_BLOCK_BURST ((uint8_t) 0x02U)
-#define DMA_CH_CTL_DT_BLOCK_BURST1 ((uint8_t) 0x03U)
-#define DMA_CH_CTL_DT_REPEAT_SINGLE ((uint8_t) 0x04U)
-#define DMA_CH_CTL_DT_REPEAT_BLOCK ((uint8_t) 0x05U)
-#define DMA_CH_CTL_DT_REPEAT_BLOCK_BURST ((uint8_t) 0x06U)
-#define DMA_CH_CTL_DT_REPEAT_BLOCK_BURST1 ((uint8_t) 0x07U)
+#define DMA_CH_CTL_DT_MASK ((uint16_t) 0x07U)
+#define DMA_CH_CTL_DT_SINGLE ((uint16_t) 0x00U)
+#define DMA_CH_CTL_DT_BLOCK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_DT_BLOCK_BURST ((uint16_t) 0x02U)
+#define DMA_CH_CTL_DT_BLOCK_BURST1 ((uint16_t) 0x03U)
+#define DMA_CH_CTL_DT_REPEAT_SINGLE ((uint16_t) 0x04U)
+#define DMA_CH_CTL_DT_REPEAT_BLOCK ((uint16_t) 0x05U)
+#define DMA_CH_CTL_DT_REPEAT_BLOCK_BURST ((uint16_t) 0x06U)
+#define DMA_CH_CTL_DT_REPEAT_BLOCK_BURST1 ((uint16_t) 0x07U)
 
 #define DMA_CH_CTL_R_DT_MASK (DMA_CH_CTL_DT_MASK << DMA_CH_CTL_R_DT_BIT)
 #define DMA_CH_CTL_R_DT_SINGLE (DMA_CH_CTL_DT_SINGLE << DMA_CH_CTL_R_DT_BIT)
@@ -184,11 +184,11 @@
 
 
 /*-----------*/
-#define DMA_CH_CTL_R_PIN1_BIT ((uint8_t) 1U)
+#define DMA_CH_CTL_R_PIN1_BIT ((uint16_t) 1U)
 
-#define DMA_CH_CTL_PIN1_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_PIN1_RISING ((uint8_t) 0x00U)
-#define DMA_CH_CTL_PIN1_FALLING ((uint8_t) 0x01U)
+#define DMA_CH_CTL_PIN1_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_PIN1_RISING ((uint16_t) 0x00U)
+#define DMA_CH_CTL_PIN1_FALLING ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_PIN1_MASK (DMA_CH_CTL_PIN1_MASK << DMA_CH_CTL_R_PIN1_BIT)
 #define DMA_CH_CTL_R_PIN1_RISING (DMA_CH_CTL_PIN1_RISING << DMA_CH_CTL_R_PIN1_BIT)
@@ -196,11 +196,11 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_PIN2_BIT ((uint8_t) 2U)
+#define DMA_CH_CTL_R_PIN2_BIT ((uint16_t) 2U)
 
-#define DMA_CH_CTL_PIN2_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_PIN2_RISING ((uint8_t) 0x00U)
-#define DMA_CH_CTL_PIN2_FALLING ((uint8_t) 0x01U)
+#define DMA_CH_CTL_PIN2_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_PIN2_RISING ((uint16_t) 0x00U)
+#define DMA_CH_CTL_PIN2_FALLING ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_PIN2_MASK (DMA_CH_CTL_PIN2_MASK << DMA_CH_CTL_R_PIN2_BIT)
 #define DMA_CH_CTL_R_PIN2_RISING (DMA_CH_CTL_PIN2_RISING << DMA_CH_CTL_R_PIN2_BIT)
@@ -208,11 +208,11 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_PIN3_BIT ((uint8_t) 3U)
+#define DMA_CH_CTL_R_PIN3_BIT ((uint16_t) 3U)
 
-#define DMA_CH_CTL_PIN3_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_PIN3_RISING ((uint8_t) 0x00U)
-#define DMA_CH_CTL_PIN3_FALLING ((uint8_t) 0x01U)
+#define DMA_CH_CTL_PIN3_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_PIN3_RISING ((uint16_t) 0x00U)
+#define DMA_CH_CTL_PIN3_FALLING ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_PIN3_MASK (DMA_CH_CTL_PIN3_MASK << DMA_CH_CTL_R_PIN3_BIT)
 #define DMA_CH_CTL_R_PIN3_RISING (DMA_CH_CTL_PIN3_RISING << DMA_CH_CTL_R_PIN3_BIT)
@@ -220,11 +220,11 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_PIN4_BIT ((uint8_t) 4U)
+#define DMA_CH_CTL_R_PIN4_BIT ((uint16_t) 4U)
 
-#define DMA_CH_CTL_PIN4_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_PIN4_RISING ((uint8_t) 0x00U)
-#define DMA_CH_CTL_PIN4_FALLING ((uint8_t) 0x01U)
+#define DMA_CH_CTL_PIN4_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_PIN4_RISING ((uint16_t) 0x00U)
+#define DMA_CH_CTL_PIN4_FALLING ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_PIN4_MASK (DMA_CH_CTL_PIN4_MASK << DMA_CH_CTL_R_PIN4_BIT)
 #define DMA_CH_CTL_R_PIN4_RISING (DMA_CH_CTL_PIN4_RISING << DMA_CH_CTL_R_PIN4_BIT)
@@ -232,11 +232,11 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_PIN5_BIT ((uint8_t) 5U)
+#define DMA_CH_CTL_R_PIN5_BIT ((uint16_t) 5U)
 
-#define DMA_CH_CTL_PIN5_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_PIN5_RISING ((uint8_t) 0x00U)
-#define DMA_CH_CTL_PIN5_FALLING ((uint8_t) 0x01U)
+#define DMA_CH_CTL_PIN5_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_PIN5_RISING ((uint16_t) 0x00U)
+#define DMA_CH_CTL_PIN5_FALLING ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_PIN5_MASK (DMA_CH_CTL_PIN5_MASK << DMA_CH_CTL_R_PIN5_BIT)
 #define DMA_CH_CTL_R_PIN5_RISING (DMA_CH_CTL_PIN5_RISING << DMA_CH_CTL_R_PIN5_BIT)
@@ -244,11 +244,11 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_PIN6_BIT ((uint8_t) 6U)
+#define DMA_CH_CTL_R_PIN6_BIT ((uint16_t) 6U)
 
-#define DMA_CH_CTL_PIN6_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_PIN6_RISING ((uint8_t) 0x00U)
-#define DMA_CH_CTL_PIN6_FALLING ((uint8_t) 0x01U)
+#define DMA_CH_CTL_PIN6_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_PIN6_RISING ((uint16_t) 0x00U)
+#define DMA_CH_CTL_PIN6_FALLING ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_PIN6_MASK (DMA_CH_CTL_PIN6_MASK << DMA_CH_CTL_R_PIN6_BIT)
 #define DMA_CH_CTL_R_PIN6_RISING (DMA_CH_CTL_PIN6_RISING << DMA_CH_CTL_R_PIN6_BIT)
@@ -256,11 +256,11 @@
 /*-----------*/
 
 /*-----------*/
-#define DMA_CH_CTL_R_PIN7_BIT ((uint8_t) 7U)
+#define DMA_CH_CTL_R_PIN7_BIT ((uint16_t) 7U)
 
-#define DMA_CH_CTL_PIN7_MASK ((uint8_t) 0x01U)
-#define DMA_CH_CTL_PIN7_RISING ((uint8_t) 0x00U)
-#define DMA_CH_CTL_PIN7_FALLING ((uint8_t) 0x01U)
+#define DMA_CH_CTL_PIN7_MASK ((uint16_t) 0x01U)
+#define DMA_CH_CTL_PIN7_RISING ((uint16_t) 0x00U)
+#define DMA_CH_CTL_PIN7_FALLING ((uint16_t) 0x01U)
 
 #define DMA_CH_CTL_R_PIN7_MASK (DMA_CH_CTL_PIN7_MASK << DMA_CH_CTL_R_PIN7_BIT)
 #define DMA_CH_CTL_R_PIN7_RISING (DMA_CH_CTL_PIN7_RISING << DMA_CH_CTL_R_PIN7_BIT)
